LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/wide-gamut-canvas/2d.color.space.p3.to.p3-expected.txt

2d.color.space.p3.to.p3
test getImageData with display-p3 and uint8 from display p3 uint8 canvas
Actual output:

FAIL test getImageData with display-p3 and uint8 from display p3 uint8 canvas assert_approx_equals: expected 62 +/- 2 but got 50
